compositor: do not request repaint in output_enable
When enabling a weston_output, the compositor damages the entire output,
to ensure that the entire output is redrawn. However, this triggers a
repaint, which might happen before anything is actually drawn. The
result is a black frame before the shell is displayed.

Damage the entire output region, but do not trigger the repaint when
enabling a weston_output.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>