clients: Add simple-hdr-video client
The client uses ffmpeg to decode videos and to extract the colorspace
information and HDR metadata. This metadata and colorspace
information is sent using the colorspace-unstable-v1-protocol and
hdr-metadata-v1-protocol to the compositor along with the buffer.
The client supports YUV420 and YUV420_10 bit formats. It converts the
YUV420_10 bit format to P010 (interleaved) format.
P010 format support was introduced in mesa from the following commit:
https://gitlab.freedesktop.org/mesa/mesa/commit/2a2e69f975bd791d64a88553d5b1d5eda7e8abdf

Signed-off-by: Harish Krupo <harish.krupo.kps@intel.com>