libinput: support high-resolution scroll wheels
The new libinput API provides three new event types, replacing the old
POINTER_AXIS event with an axis source attached. The new event types
POINTER_SCROLL_FINGER/WHEEL/CONTINUOUS have that source encoded in the
event type and, in the case of POINTER_SCROLL_WHEEL, returns scroll wheels
as fractions of 120 per logical click (that's basically the windows API).
The old event type can now be ignored where the new one is processed,
this patch ifdefs out the various bits depending on the libinput
version.

We use this to switch weston over to use 120 instead of the old-style
discrete events. Where the new libinput isn't available, we just use the
discrete values and multiply them by 120.

On the client side, where the client doesn't support 120 values, we
divide again to get the equivalent discrete value. discrete and v120 are
mutually exclusive on the wayland client side.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: José Expósito <jose.exposito89@gmail.com>