compositor-drm: Support KMS crtc out-fences
The out-fence for each output is temporarily stored in drm_output_state
after a successful atomic commit. It is kept valid for just long enough,
until drm_output_assign_state, so that it can be used as the release
fence in an upcoming commit.

This reduced out-fence lifetime was chosen to simplify code, but may
need to be expanded if in the future we need to use out-fences for
different purposes.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>