input: Add support for zwp_text_input_v3 and zwp_input_method_v2
Some clients such as gtk+3 require the zwp_text_input_v3 protocol
for input-method handling. This patch adds this protocol support
to text-backend and also modifies weston-keyboard to support
zwp_input_method_v2, which is an update for the v1 protocol and aligned
with the updated text-input protocol.

The patch also adds support for zwp_virtual_keyboard_v1, used to emulate
the behaviour of a physical keyboard.

Note that the updated input-method v2 protocol does not support the
input-panel interface as it did on v1. The input panel interface is
implemented by desktop-shell to properly display the virtual keyboard
interface.

To avoid depending on both input method v2 and v1 (for input-panel only),
this patch adds a new weston-input-panel internal protocol (a copy of
the old interface from input-method v1), which is then implemented by
desktop-shell and used by weston-keyboard when displaying its window.

Signed-off-by: Andre Moreira Magalhaes (andrunko) <andre.magalhaes@collabora.co.uk>
Signed-off-by: Silvan Jegen <s.jegen@gmail.com>