gl-renderer: use intermediate texture for linear light blending
Linear light blending is required when the presented surfaces are
of different colorspaces. The surfaces are linearized, colorspace
convertion is applied (if required) and then blended together. All
this is done to an intermediate texture so that it can be blitted
on to the framebuffer while applying the required non linear curve.

v2: Use /* */ instead of // (Pekka)
    Rename fbo and tex to shadow_{fbo,tex} (Pekka)
    Check for OpenGLES capabilities before creating
    shadow_{tex,fbo} (Pekka)

Signed-off-by: Harish Krupo <harishkrupo@gmail.com>