compositor: Add timing debug scope
Add a 'timing' scope which prints details which can be useful for
debugging presentation timing issues.

This detail was useful in debugging insane calculated repaint delay
messages with drm-imx driver, and may be useful for other timing
debugging.

Currently it contains only the details I needed to debug that particular
issue, but others could be added to it across all backends and the
compositor core.

Signed-off-by: Robert Beckett <bob.beckett@collabora.com>