protocol/content-protection: Add new DC_ONLY, HDCP protection level
This is an new CP-level, not available in the HDCP spec.

Instructs the compositor to censor the surface's content on a non-secure
(not HDCP-capable) output when there's an active recording, similar to
what HDCP does.

This content protection level is useful in rather closed setup systems
where HDCP is not available.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>