libweston: Schedule an idle task to redraw the output when censorship stops
This is needed as to avoid displaying the placeholder graphics once the
the recording/screenshooting stops. The client might not redraw the
contents of the view and as such the placeholder can still be be seen.

This also happens when screenshooting/recording stops for enforced
protection modes on secure-outputs. so it is necessary for that as well.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Suggested-by: Pekka Paalanen <pekka.paalanen@collabora.com>