renderer-gl: Improve screenshot implementation
Added new screenshot APIs which will copy surface/output content to
a provided DMA buffer file descriptor.
This is done by creating an egl image using the DMA buffer fd and set
it as a target to a texture. Then we will create and attach a
framebuffer to this texture and copy the surface/output content to the
framebuffer (in turn to the texture) using glDrawArrays.

Signed-off-by: K J Rajendraprasad <KarammelJayakumar.Rajendraprasad@in.bosch.com>