clients: Improve screenshot implementation
The current screenshot implementation is less performant as it
is not H/W accelerated. One way to improve the screenshot
performance is to use DMA buffers. The idea is that the weston
screenshooter client will allocate a DMA buffer and sent it to
the compositor. The compositor then copies the surface/output
content to the DMA buffer and returns it. The client will then
copy the DMA buffer contents to a file.

Signed-off-by: K J Rajendraprasad <KarammelJayakumar.Rajendraprasad@in.bosch.com>