gl-renderer: Add FBO path for scaled screenshots
When taking screenshots of a renderer-follows-scale output, we still
need to provide a screenshot the full size of the output, like we would
do with a scale on final display.

Unfortunately this is difficult due to the way that screenshots are
handled: they are implemented by calling a read_pixels hook after a
repaint. The fact that we are going through repaint means that we need
to have valid output content inside the EGLSurface. We cannot texture
from our EGLSurface backbuffer, only use ReadPixels to pull the content
back into CPU memory.

This takes a maximally-pessimal approach: we render the output to the
EGLSurface as normal, call glReadPixels to pull the surface content back
into CPU memory, upload it straight back into a new GPU texture, then
use the GPU to do a stretch blit into a scratch FBO which we again call
glReadPixels on.

Doing better would require us to know that a screenshot would be
happening at repaint time, in which case we would render to an FBO from
the start, blit from the FBO to the EGLSurface for presentation, blit
from the first FBO to another FBO for scaling, and then ReadPixels back
from the second FBO. This doesn't seem like a big enough win to
implement right now.

Signed-off-by: Daniel Stone <daniels@collabora.com>