fullscreen-shell: do not take sub-surfaces into account on present_surface
Surface positioning was affected by the position and size of the sub-surfaces.
If subsurfaces were added or changed, the positioning of the main surface did
not change immediately. But with a new present_surface, the positioning was
changed in an unexpected way.

The intention of this patch is a consistent and comprehensible influence of the
sub-surfaces. Regardless of whether the sub-surfaces were added before or after
present_surface. By no longer taking the sub-surfaces into account in positioning
the main surface, the code becomes simpler and more understandable.

This problem was discussed and documented with logs and screenshots
at #387 https://gitlab.freedesktop.org/wayland/weston/-/issues/387.

Signed-off-by: ahe <Andreas.Heynig@meetwise.com>