drm: Add start-up delay until a graphics device is present
The initialization of a graphics driver could be slower than Weston,
leading Weston to fail.

So, a wait and probe loop is added to wait until any graphics be present,
and then, proceed with DRM-backend initialization. If no graphics card is
available in the timeout frame window, Weston will fail as expected.

That is made by adding a loop that sleeps the process for 10ms and probes
the udev/DRM. Until either, a card to be found or 20 loops succeed(0.2s timeout).

Closes: #173
Signed-off-by: Igor Matheus Andrade Torrente <igortorrente@outlook.com>