data-device: avoid duplicating data offer for the same client
`weston_data_source_send_offer` is called to create a new data offer every time
when the seat focuses a surface. This patch will evite create a new data offer
for the same client for the same source, it fixes the bug of #376, where the
data device duplicates data offers thus it never been pasted to firefox.

Signed-off-by: xichen zhou <sichem.zh@gmail.com>