doc/systemd: system service example
There are many bad and even worse attempts to make Weston run as a
systemd service, and very few good ones. Let's add a good one as an
example in upstream: does not use weston-launch, does not run weston as
root, does not need wrapper scripts, and relies on logind and PAM.

This example has been composed from a couple of real-world Weston unit
files used in production.

The service file is not installed by the build. It would likely need
small adjustments for each distribution or system to be deplyed on.

Fixes: https://gitlab.freedesktop.org/wayland/weston/-/issues/7
Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>
Signed-off-by: Stefan Agner <stefan@agner.ch>