weston-rotator: Add weston-rotator protocol
Protocol and server implementation. This is a follow-up continuation
work done by Emmanuel Gil Peyrot (<linkmauve@linkmauve.fr>) at
[1]. It re-uses the same design, the server implementation being done
as a weston plug-in and it is hooked to start-up with desktop-shell. It
allows to re-position/change the orientation the output dynamically
based on a wl_output.transform value passed on by the client. The client
is the one that decides the new transform value, allowing to read the
data from a sensor or do it manually.

Some small changes have been further added:

- 'done' event when orientation was completed. Client can check the
  transform value against the one that it wanted initially. The
  transform value could also be retrieved by the client if it is adding
  listeners for wl_output.geometry event
- either a wl_output or a string can be passed when doing an orientation
  change, giving the client the ability to modify the orientation from
  the command line, but only if weston was started with the debug option.
  The string on the other hand, is used to specify a sensor (accelerometer) name.
- re-uses wl_output.transform instead of hard-coding ones

[1] https://lists.freedesktop.org/archives/wayland-devel/2016-August/030396.html

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>