desktop-shell: Plug-in the suspend/unsuspend output mechanism
We have now everything in place to install listeners for receiving
ack_configure requests, and with it to to suspend the output, while
waiting for all full screen or maximized views to send back the
ack_configure requests.

We keep track of shell_surfaces that still haven't received the
new size and remove one by one until all of them have acknowledged the
the new size.

We also install a timer which will unfreeze the output in case
any of the application will deliberately not send the ack_configure
request.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>