libweston: Add helpers to suspend/resume a weston_output
And with it, some timer helpers to allow unfreezing the output. This
small API can be used to increase a reference count over the suspendend
output and can be used to when changing the orientation of the output.

The repaint machinery will skip repainting as long as there's at
least a reference count set for it, and it  is necessary to avoid
rendering glitches while the client re-configures its surface width
and height.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>