desktop-shell: Handle all other shell_surfaces
This patch plug-ins initial support to handle regular xdg-shell
surfaces, in the output resize signal, and pass the new width and height
which is being swapped when a 90 degree orientation takes place. That
happens only for fullscreen and maximized views, while for others we try
to reset their position to match the new orientation as to avoid placing
views outside of the output viewing area. We only restore the positon when
going back to initial orientation.

The output_resize signal handler only dealt so far with panels and background.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>