libweston, desktop-shell: Add a method to flush out frame callbacks
In this manner we send out frame callbacks from other call sites, in our
case when sending out the configure event for fullscreen and maximized
views. This is necessary to avoid clients waiting for frame callback
being stuck and unable to process configure events sent
previously/earlier.

This also is particularly problematic as the compositor will wait for
the ack_configure requests and mitigated with the help of the time out
timer. Nevertheless, sending the frame callbacks gives the client the
opportunity to handle out the configure events.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>