shared: Add shared transform file
And migrate transform related functions to it. Also, make use of it when
displaying the scene-graph.

We're going to use the transform functions in the client, and migrating
this to libweston would imply a link to it, so it seems much more
natural to have it as a shared bit. It also avoids duplicating the same
thing in the client.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>