backend-drm: Use dumb buffer for cursor buffer
Fixes: #375

Currently the DRM-backend depends on GBM for cursor
planes, more specifically for the buffers for cursor
planes. This means that cursor planes can be used only
if GBM is initialized, which means using GL-renderer.
Hence when using Pixman-renderer, cursor plane is not used.

This changes adds support to use dumb buffer as cursor buffer
for cursor plane and allow it to be used with both pixman
and gl renderer.This also avoids the cost of repainting
for only cursor motion with pixman renderer.

GBM path is completely removed for cursor related stuff.

Signed-off-by: Veeresh Kadasani <veeresh.kadasani@huawei.com>