tests, meson_options: Add test-dev-install option
As the test-plugin.so can be seen as back-door into the compositor
internals, only expose it as a build time feature, but still making it
easier for people to use the test harness in case they would like to do
so. Add a short explanation into README to inform users of that.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>