kiosk-shell: Force a surface geometry change for xwayland surfaces
This would allow re-building of the surfaces state list to make sure
that the mapping of the xwayland surface would actually trigger a
surface transform signal, and with it, signal the correct positioning of
a xwayland surface, with the help of signal handler installed for that
transform_signal event.

For some xwayland surfaces, not doing an explicit surface geometry
change, would result in a incorrect positioning of the xwayland surface
when it is started as fullscreen, on a another output.  As such this
would only be noticeable in a multiple outputs case.

The issue is only seen on kiosk-shell due to the fact as on other shells,
the desktop one, a transition to other states (maximized, to fullscreen,
to a normal one) would actually involve a surface geometry change.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>