tests: Introduce X11 client helper
Helpers to create and manipulate X11 clients, to perform and handle
various transition states, to retrieve atoms, send events, allowing to
set up simpler and complex uses-cases.

This patch creates a static library for helpers which X11 clients can
link against.

With it, redoes the xwayland test to make use of it, and we remove Xlib
and instead use XCB.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>