xwayland-states-test: Added basic xwayland tests
This tests out map, maximized and fullscreen working states. The
tests will start with the xwayland surface as maximized/fullscreen then go
back to their initial/normal state.

While this can also be done with using a client message (toggle
function) we need to be able to start, from the beginning the xwayland
surface as max/fullscreen to be able to exercise some of the code in XWM.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>