ivi-shell: damage all outputs after initializing shell
If the compositor does not damage newly enabled outputs, the ivi_layout_internal
runs into a timeout, because there are no events, that would help the main loop
to continue and exit.

Damaging all outputs during the ivi-shell initialization hides this issue,
because the idle repaint handler causes an idle event and helps to finish the
main loop.

Signed-off-by: Michael Tretter <michael.tretter@posteo.net>