protocol/weston-test: Add a new interface xwayland_test_xwayland
This patch expands the weston_test protocol to include
xwayland_test_xwayland interface which helps out to keep track of the
repaint events coming through Xwayland.

Introduces a request that adds a listener for the repaint events and a
couple of helpers that are used to block/wait until all re-paint events
have been accounted for.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>