backend-drm: Add support for handling release_fence
If DRM_CAP_RELEASE_FENCE capability is exposed by KMS, then
include the RELEASE_FENCE_PTR as part of the atomic commit.
After getting a valid release fence fd, take an additional reference
on all the fbs included in the atomic commit and also register
a handler for the release fence fd by adding it to the main event loop.

This patch in addition to other related DRM driver patches can solve:
https://gitlab.freedesktop.org/wayland/weston/-/issues/514

Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>