input: Add keyboard grab protocol implementation
This makes it possible to send a input_method commit string request when
a key is pressed on the non-virtual keyboard. That way we can reset the
preedit string on a (real) keyboard key press and not only when we press
a key on the virtual on-screen keyboard.

Note that this code is very heavily inspired by Sway's implementation
of the protocol.

Signed-off-by: Silvan Jegen <s.jegen@gmail.com>