window.c: use xdg-decoration-unstable-v1
This commit uses the xdg-decoration-unstable-v1 protocol in toytoolkit,
to communicate the toolkit's preferred way to render window decorations,
and to hide the frame decorations if the compositor insists on server-side
decoration rendering. This can make Weston's client programs fit in better
under other compositors which support xdg-decoration.

The preferred decoration mode is set to client-side, so that client behavior
stays close to that under Weston, which does not yet support xdg-decoration.

Signed-off-by: Manuel Stoeckl <code@mstoeckl.com>