libweston: Add y_flip to weston_surface_copy_content
Until now, weston_surface_copy_content() has been mostly just a
debugging API, with an austere feature set.

The RDP backend intends to add support for RDP-RAIL, which allows
remoting of individual client windows (rootless operation). It's
useful for that backend to use this API to fetch surface content.

In RDP-RAIL, the client expects the cursor content to be Y inverted,
so we add a y-flip parameter.

Co-authored-by: Steve Pronovost <spronovo@microsoft.com>
Co-authored-by: Brenton DeGeer <brdegeer@microsoft.com>
Signed-off-by: Hideyuki Nagase <hideyukn@microsoft.com>
Signed-off-by: Steve Pronovost <spronovo@microsoft.com>
Signed-off-by: Brenton DeGeer <brdegeer@microsoft.com>