libweston: Allow scaling in weston_surface_copy_content
If a target height or width is specified, the content will be scaled
accordingly. A 0 for height or width results in using the source size
for that dimension. This is so we can present scaled data to an
RDP-RAIL client.

Note: For now, the GL implementation asserts if scaling is required.

Co-authored-by: Steve Pronovost <spronovo@microsoft.com>
Co-authored-by: Brenton DeGeer <brdegeer@microsoft.com>
Signed-off-by: Hideyuki Nagase <hideyukn@microsoft.com>
Signed-off-by: Steve Pronovost <spronovo@microsoft.com>
Signed-off-by: Brenton DeGeer <brdegeer@microsoft.com>