compositor: fix crash in xdg_output_manager_get_xdg_output after head removal
When a head is removed and destroyed, the associated wl_output resources user
data is reset to null.

When a client calls zxdg_output_manager_v1::get_xdg_output on a removed
wl_output (which the client might not be aware of yet), this could trigger a
crash as the output resource would be assumed to still represent a valid head.

Signed-off-by: Arnaud Vrac <avrac@freebox.fr>