backend-drm: Use the seat name over cmd line
The seat name passed over the command line are used to identify the DRM
device. Yet, the DRM plug-in API ignores it entirely and relies only on
the configuration file to specify a seat.

When the configuration file doesn't contain any seat explicitly
specified, we'll default to an empty string, and eventually end up with
no seats in the system. That happens even if we would supply a valid
seat, with inputs attached to it.

Adjust that by storing the seat_id and take it into consideration if
specified over the command line.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>