backend-drm: Create a fake seat if none specified
Starting with commit 'backend-drm: Use the seat name over cmd line' we
would be able to create a fake-seat, by just specifying an invalid
seat name.

In the same time, in a system where there are actually no input devices
and no one specifies an explicit seat (as in to create a fake one) we
would end with no seat available in the system.

This makes things a bit easier for situations where users would just
specify a DRM node, with input devices being accessed directly, or
rather with no input devices whatsoever.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>