clients/presentation-shm: add RUN_MODE_NO_THROTTLE
In this mode we don't have any throttling mechanism, we commit lots of
SHM buffers, non-stop.

This is useful to test if compositors are discarding the content of
these commits instead of right off sending them to composition. Weston
for instance does not upload these buffers until a certain point before
the composition. Compositors that are not doing that may freeze with
such clients.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>