tests: add DRM/KMS composite bypass test
This introduces a composite-bypass test, in order to exercise Weston's
planes management, among other mechanisms.

It creates buffers that can be given to the primary plane of the
graphics card so they can be displayed directly, bypassing the
renderers. In order to achieve this, we need hardware buffers.

As we want to run this in our CI using VKMS, we use VGEM in order to
create "hardware buffers" in the absence of a real graphics card. Then
these buffers can be exported to VKMS by using the linux-dmabuf API.

As it skips rendering, it depend on the writeback screenshooter in order
to validate if everything went well.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>