tests/drm-writeback-screenshot-test: Grab the first available output
This is basically a small follow-up from commit
'drm-smoke-test: Explicitly set-up the output name DRM test' which did
the same thing.

With multiple virtual outputs in the system, we would override the output and
pick/use up the last one available, with the client placing the
window/surface on the first output. This would make things easier when
testing and transitioning to multiple virtual outputs.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>