backend-drm: add get_modes function to backend API
The get_modes() call allows the frontend to ask the backend for the supported
modes. The frontend may print the existing modes (including the full mode line)
and be able to select a specific mode from the available modes.

Furthermore, the frontend is able to detect that an output doesn't expose any
modes and may handle this situation.

After the frontend has called get_modes(), a monitor may be unplugged, which
frees the detected modes on the output. To prevent that the frontend keeps
dangling references to the detected modes, the mode lines are passed as copy to
the frontend.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>