backend-drm: wrap drmModeModeInfo into struct weston_drm_modeline
Wrapping the drmModeModeInfo into struct weston_drm_modeline allows to attach
further information to the mode info. This allows handling the different formats
that can be specified as a modeline in weston.ini in an explicit struct, and
finally moving the format parsing to the frontend.

Introduce a new Weston specific struct drm_modeline to represent a modeline,
which can be parsed from the string formats defined in weston-drm(7).

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>