kiosk-shell: bind key to weston surface
weston surface displays on fullscreen mode with kiosk shell,
multiple surfaces overlap each other when running client
applications on the same output monitor. This is to bind a
key to an application's weston surface, then it's able to
display this specific surface on monitor top level via
sending an activate keyboard event to it.

Signed-off-by: Sophia Gong <lili.gong@amd.com>