pixel-formats: Use all three GL format types
GL comes with three sets of format enums: 'internal' formats, which can
either be sized or unsized, as well as 'format' to describe the colour
channels and 'type' to describe the layout of the data.

Previously we always used format + type, because that's the only thing
GLES2 allows. However, GLES3 allows us to use sized internal formats
instead, so let's start doing that in order to be more explicit.

Signed-off-by: Daniel Stone <daniels@collabora.com>