gl-renderer: Use internal formats in TexImage2D
GLESv3 allows us to use explicitly-sized internal formats when we upload
pixel data, rather than being inferred from a base format and type. Use
this where we can (with a small get-out because not all the pixel
formats have internal formats declared yet ...), with a fallback to base
formats for GLESv2.

Signed-off-by: Daniel Stone <daniels@collabora.com>