gl-renderer: Use glTexStorage2D for SHM buffers
glTexImage2D does two things: it uploads pixel data to the texture, but
it can also respecify any element of the texture, changing its
dimensions/format/etc.

Since these are already immutable for us, use glTexStorage2D up front
when creating SHM buffers. This means that we have to use
glTexSubImage2D instead of glTexImage2D to upload - as calling
glTexImage2D on an immutable buffer is not allowed - but this shouldn't
make any difference.

Signed-off-by: Daniel Stone <daniels@collabora.com>