gl-renderer: Allow GL_BGRA for renderbuffers
Previously we were declaring the internalformat of GL_BGRA to actually
be GL_RGBA, because that was the thing we had which worked. Now that
we've worked through the spec issues and Mesa bugs, we can actually use
GL_BGRA_EXT and GL_BGRA8_EXT as internal formats for renderbuffers.

Signed-off-by: Daniel Stone <daniels@collabora.com>