libweston: move weston-assert from shared to include/libweston
In the next commits we'll integrate the weston-assert macros with our
log infrastructure. In order to do that, we'll restrict such macros to
libweston and its users.

In the future we plan to add another header of macros specifically for
the test suite and clients. The difference is that this one won't take
the *compositor parameter that we currently have in the weston-assert
macro signatures, and they will log directly to stderr.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>