libweston: detach writeback connectors when dpms switch to off
When DPMS switches to off, Weston will commit a new state to disable the
currently used CRTCs. If there are any connectors connected to these CRTC,
the kernel will be unable to commit the new state and throw an error.
After taking a screenshot, the writeback connector is still attached
to the CRTC. This commit will detach the writeback connector from the
CRTC when DPMS switches to off. Create new pointer in crtc structure,
this pointer will point to writeback connector ojbect connected to this
crtc. And re-attach them when dpms switch back to on.

Signed-off-by: Chien Phung Van <chien.phungvan@vn.bosch.com>