desktop-shell: shrink windows not to be obscured by input panel
This introduces the boolean option input-panel-shrink in the
shell section of weston.ini.

When this option is set and an input panel becomes
visible, the desktop shell checks whether any windows on
screen become overlapped by the input panel. If this is
the case, the heights of the affected windows are shrunk until
they stop overlapping. Maximized and fullscreen windows
are handled by removing their maximized or fullscreen
status while the input panel is visble.

When the input panel is hidden, all affected windows
are restored to their original states.

In addition a slide-out animation is added to the input
panel when it is hidden to complement the existing slide-in
animation when it appears.

Signed-off-by: Sebastian Urban <surban@surban.net>