clients: client verifies the screenshot feature using the writeback source
Create a Weston client using the writeback source to take a screenshot.
This application will generate an image, request a fullscreen surface
to display that image, and then send a request to Weston to take a
screenshot. Once the screenshot is successfully captured, it will compare
the content of the shot image with the reference image. They will be saved
as PNG files if there are any differences in content. By default, DMA
buffer will be used, and if there are any errors with the DMA buffer,
SHM will be used. Alternatively, users can specify the buffer type or
DRM device through arguments. They can also use the '-h|--help' flag
to get more detailed information.

Signed-off-by: Chien Phung Van <chien.phungvan@vn.bosch.com>