libweston: Add frame rate support
This is a continuos of work from
https://gitlab.freedesktop.org/wayland/weston/-/merge_requests/24

Adds frame rate counters to the scene-graph debug scope. The timer
interval is set by default to 3 seconds. As the debug scope is a one
time shot type of scope, we use a timer to periodically compute the
current commits in that interval as as the frame rate per interval.
Solely based on wl_surface::commit call.

This is helpful to identify what frame rates clients achieve to debug
performace issues as to provide an objective common group. Doing this as
the compositor level helps cases when GL is not in use. Using the
debug-scope aggreates all surfaces at the same place.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Signed-off-by: Changwoo Cho <changwoo.cho@lge.com>