gitlab-ci.yml: Add dedicated stage and container for docs deploy
This creates new container, that will only be created on main, to allow
building the docs and push/deploy docs to in pages stage.

This means that for main we'll have 3 jobs running, container one,
build / docs, and deploy, distinct from the MR context, which would run
also this for x86-64, but all others jobs as well.

It doesn't avoid running any jobs but should be considerably faster than
running all of them. Probably caching the x86-64 jobs would help and
could be something we could look into next.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>