color: add function to craft color curves from color parameters
When we create a parametric color profile (we still do not fully
support, but plan to), a struct weston_color_tf_info is created.

Although we support a few color curves on Weston, we don't have
anything mapping from the tf_info object to such color curves. This
patch adds a function to do that.

That should be helpful to create color transformations that contain
parametric color profiles.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>