backend-drm: Use plane's zpos_min to check for underlay ability
Previously, whether a KMS plane is underlay-able is determined by
whether it's zpos_max is < the primary plane's zpos_min. In other words,
a plane will only be considered underlay-able if its entire valid zpos
range is under the primary plane's lowest zpos.

This is too restrictive - it's possible for planes to have a valid zpos
range that spans below and above the primary's zpos range.

Therefore, allow planes to be used as underlays if their zpos_min is <
the primary plane's zpos_min.

Because a plane can now - if supported - be used as an underlay or an
overlay, rename the plane's type name to "secondary". Then, print it's
underlay/overlay assignment once a decision is made.

Signed-off-by: Leo Li <sunpeng.li@amd.com>