frontend: allow to create automatic param color profiles from .ini
Add a new option to "color_profile" key on output section of weston.ini:
automatic profiles.

Automatic color profiles are created accordingly to the colorimetry mode
and the EOTF that the output is using. We use CTA-861 for the parameters
as much as possible. There are also optional flags, and in such case we
also take into consideration the EDID data.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>