color: add get_output_to_blend_lut()
With this function, we can get the blend-to-output color transformation
in 3x1D LUT's.

The DRM-backend will make use of that to off-load the blend-to-output
color transformations, what happens in the next commits.

Note that this may result in precision issues when such transformation
contains an OETF whose section of the curve in which the pixels are less
bright is too steep. This combined with the fact that humans perceive
more subtle luminance changes in the dark may produce the undesired
effect of color banding.

But as, for now, the DRM-backend has support only for LUT's for
post-blending transformations, we may use that temporarily.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>